requirements.txt
playwright
google-generativeai